update: modify 'dutycycle' from float type to integer

+++ src/motor/src/bldc.h
@@ -12,7 +12,7 @@ public:
     virtual int get_erpm() const;
     virtual int get_e_period() const;
     virtual int get_current() const;
-    virtual void set_throttle(float v);
+    virtual void set_throttle(int v);
     virtual void stop();
     virtual void beep(uint32_t freq, VolumeLevel volume);
     virtual void arm(bool state);
